backend/flask_api/requirements.txt
openai-whisper
cohere
Flask
Flask-CORS